+++ packages/gamut/src/DataList/DataGrid.tsx
@@ -1,3 +1,5 @@
+import { system } from '@codecademy/gamut-styles';
+import { StyleProps } from '@codecademy/variance';
 import { ComponentProps, useMemo } from 'react';
 
 import { List } from '../List';
@@ -14,7 +16,10 @@ export interface DataGridProps<
   Cols extends ColumnConfig<Row>[]
 > extends MarshaledColProps,
     DataListControls<Row, IdKey, Cols>,
-    Omit<ComponentProps<typeof List>, 'header' | 'id'> {
+    Omit<
+      ComponentProps<typeof List>,
+      'header' | 'id' | 'rowBreakpoint' | keyof StyleProps<typeof system.space>
+    > {
   /** Whether the data inside is loading and should be indicated to the user somehow */
   loading?: boolean;
   /** Whether an additional header element should be added */
@@ -49,7 +54,7 @@ export function DataGrid<
     emptyMessage,
     hideSelectAll = false,
     scrollToTopOnUpdate = false,
-    ...rest
+    id,
   } = props;
 
   const empty = rows.length === 0;
@@ -96,7 +101,6 @@ export function DataGrid<
     <ListStateContext.Provider value={{ query }}>
       <ListControlContext.Provider value={listControls}>
         <List
-          {...rest}
           as="table"
           emptyMessage={emptyMessage ?? <EmptyRows />}
           header={
@@ -110,6 +114,8 @@ export function DataGrid<
             ) : null
           }
           height={height}
+          id={id}
+          loading={loading}
           minHeight={minHeight}
           scrollToTopOnUpdate={scrollToTopOnUpdate}
           scrollable={scrollable}