+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.54.6](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.5...@codecademy/gamut-illustrations@0.54.6) (2025-07-17)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
 ### [0.54.5](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.4...@codecademy/gamut-illustrations@0.54.5) (2025-06-17)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations