+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.523](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.522...@codecademy/gamut-kit@0.6.523) (2025-07-17)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.522](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.521...@codecademy/gamut-kit@0.6.522) (2025-07-16)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
 ### [0.6.521](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.520...@codecademy/gamut-kit@0.6.521) (2025-07-11)
 
 **Note:** Version bump only for package @codecademy/gamut-kit